backend/requirements.txt
Django==4.2.28
PyMySQL==1.1.2
python-dotenv==1.2.1
cryptography==46.0.5
djangorestframework==3.16.1
